+++ skrub/tests/test_total_std_norm.py
@@ -4,7 +4,7 @@ from numpy.testing import assert_array_almost_equal
 
 import skrub._dataframe as sbd
 from skrub import GapEncoder, StringEncoder
-from skrub._total_variance_norm import total_variance_norm
+from skrub._total_std_norm import total_std_norm
 
 try:
     import sentence_transformers  # noqa: F401
@@ -17,10 +17,10 @@ except ImportError:
 def test_normalization_invariance():
     """Normalizing twice leads to a unit norm."""
     X = np.random.randn(10, 3)
-    norm_1 = total_variance_norm(X)
+    norm_1 = total_std_norm(X)
     X_norm_1 = X / norm_1
 
-    norm_2 = total_variance_norm(X_norm_1)
+    norm_2 = total_std_norm(X_norm_1)
     X_norm_2 = X_norm_1 / norm_2
 
     assert_array_almost_equal(X_norm_1, X_norm_2)
@@ -33,15 +33,15 @@ def test_nonfinite():
     column-wise.
     """
     X = np.array([[None, np.inf, np.nan], [1, 2, 3], [4, 5, 6]])
-    assert total_variance_norm(X) == total_variance_norm(X[1:])
+    assert total_std_norm(X) == total_std_norm(X[1:])
 
 
 def test_dataframe(df_module):
     X = df_module.example_dataframe[
         ["int-col", "int-not-null-col", "float-col", "bool-col", "bool-not-null-col"]
     ]
-    norm = total_variance_norm(X)
-    assert total_variance_norm(X / norm) == pytest.approx(1)
+    norm = total_std_norm(X)
+    assert total_std_norm(X / norm) == pytest.approx(1)
 
 
 @pytest.mark.parametrize(
@@ -49,31 +49,25 @@ def test_dataframe(df_module):
     [
         StringEncoder(n_components=2),
         GapEncoder(n_components=2),
-        # pytest.param(
-        #     TextEncoder(n_components=2),
-        #     marks=pytest.mark.skipif(
-        #         not TRANSFORMERS, reason="sentence-transformers is not installed"
-        #     ),
-        # ),
     ],
 )
 def test_encoders(df_module, encoder):
     X = df_module.example_dataframe["str-col"]
     X_t = encoder.fit_transform(X)
-    assert total_variance_norm(np.array(X_t)) == pytest.approx(1)
+    assert total_std_norm(np.array(X_t)) == pytest.approx(1)
 
 
 def test_non_num_values(df_module):
     with pytest.raises(ValueError, match="must only have numeric columns"):
-        total_variance_norm(df_module.example_dataframe)
+        total_std_norm(df_module.example_dataframe)
 
     with pytest.raises(ValueError, match="must only have numeric values"):
-        total_variance_norm(np.array([[1, "a"]]))
+        total_std_norm(np.array([[1, "a"]]))
 
     with pytest.raises(
         TypeError, match="must be a Pandas, Polars dataframe or a Numpy"
     ):
-        total_variance_norm([1, 3])
+        total_std_norm([1, 3])
 
 
 def test_partial_fit(df_module):
@@ -87,4 +81,4 @@ def test_partial_fit(df_module):
     gap.partial_fit(X2)
     Xt = gap.transform(X_full)
 
-    assert total_variance_norm(Xt) == pytest.approx(1)
+    assert total_std_norm(Xt) == pytest.approx(1)